setup fuzzing for makura lib - fix b45 multiplication overflow panic - move decoder validator functions to decoder/base<x> module files
