+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["makura", "makura_cli"]
+members = ["makura", "makura_cli", "makura/fuzz"]
 resolver = "3"
